pkg-requirements.txt
adal==1.2.1
asn1crypto==0.24.0
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
elasticsearch==7.0.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
idna==2.8
isodate==0.6.0
msrest==0.6.6
msrestazure==0.6.0
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyJWT==1.7.1
pymongo==3.8.0
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
schedule==0.6.0
six==1.12.0
slackclient==1.3.1
uritemplate==3.0.0
urllib3==1.24.3
websocket-client==0.54.0
